@truncs 2015-09-25
@willsalz yes, I squashed the commit. Should I land ?