@willsalz 2015-09-25
Land it!

> On Sep 25, 2015, at 4:21 AM, Aditya Sarawgi <notifications@github.com> wrote:
> 
> @willsalz <https://github.com/willsalz> yes, I squashed the commit. Should I land ?
> 
> —
> Reply to this email directly or view it on GitHub <https://github.com/uber/tchannel-java/pull/76#issuecomment-143190342>.
>